requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
autobahn==18.5.2
Automat==0.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
dateparser==0.7.0
dnspython==1.15.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
numpy==1.14.5
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pymongo==3.6.1
pyOpenSSL==18.0.0
python-binance==0.6.9
python-dateutil==2.7.3
pytz==2018.4
regex==2018.2.21
requests==2.18.4
service-identity==17.0.0
six==1.11.0
Twisted==18.4.0
txaio==2.10.0
tzlocal==1.5.1
urllib3==1.22
zope.interface==4.5.0
